Remove FormatSEReBIC.py
Move the code next to the base class ready to combine them
